requirements.txt
asgiref==3.5.2
Django==4.1
django-autoslug==1.9.8
django-cart==1.0.4
django-crispy==1.8.1
django-crispy-forms==1.14.0
django-heroku==0.3.1
django-shopping-cart==0.1
djangorestframework==3.13.1
gunicorn==20.1.0
Pillow==9.2.0
pytz==2022.2.1
sqlparse==0.2.4
